Correção para não permitir enviar um pedido já faturado para o WMS

+++ library/Wms/WebService/Expedicao.php
@@ -530,11 +530,21 @@ class Wms_WebService_Expedicao extends Wms_WebService
         foreach ($pedidos as $pedido) {
             $PedidoEntity = $PedidoRepo->find($pedido['codPedido']);
             if ($PedidoEntity != null) {
-                if ( count($EtiquetaRepo->getEtiquetasByPedido($pedido['codPedido'], EtiquetaSeparacao::STATUS_PENDENTE_CORTE)) > 0) {
-                    throw new Exception("Pedido $pedido[codPedido] tem etiquetas pendentes de corte");
+                $statusExpedicao = $PedidoEntity->getCarga()->getExpedicao()->getStatus()->getId();
+                if (($statusExpedicao == Expedicao::STATUS_CANCELADO)||
+                    ($statusExpedicao == Expedicao::STATUS_EM_SEPARACAO) ||
+                    ($statusExpedicao == Expedicao::STATUS_INTEGRADO) ||
+                    ($statusExpedicao == Expedicao::STATUS_EM_CONFERENCIA)){
+                    if ( count($EtiquetaRepo->getEtiquetasByPedido($pedido['codPedido'], EtiquetaSeparacao::STATUS_PENDENTE_CORTE)) > 0) {
+                        throw new Exception("Pedido $pedido[codPedido] tem etiquetas pendentes de corte");
+                    } else {
+                        $PedidoRepo->remove($PedidoEntity);
+                    }
                 } else {
-                    $PedidoRepo->remove($PedidoEntity);
+                    $statusEntity           = $this->_em->getReference('wms:Util\Sigla', $statusExpedicao);
+                    throw new Exception("Pedido " . $pedido['codPedido'] . " se encontra " . strtolower( $statusEntity->getSigla()));
                 }
+                throw new Exception();
             }
         }
